requirements.txt
asgiref==3.5.0
celery==5.2.6
Django==3.2.12
django-celery-beat==2.2.1
django-celery-results==2.3.0
django-mptt==0.13.4
django-seed==0.3.1
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
psycopg2==2.9.3
psycopg2-binary
redis